requirements.txt
opencv-python>=4.9.0.80
diffusers==0.31.0
peft==0.17.0
transformers>=4.49.0
tokenizers>=0.20.3
accelerate>=1.1.1
tqdm
imageio
easydict
ftfy
dashscope
imageio-ffmpeg
wandb
omegaconf
einops
av==13.1.0
opencv-python
open_clip_torch
starlette
pycocotools
lmdb
matplotlib
sentencepiece
pydantic==2.10.6
scikit-image
huggingface_hub[cli]
dominate
nvidia-pyindex
nvidia-tensorrt
pycuda
onnx
onnxruntime
onnxscript
onnxconverter_common
flask
flask-socketio
torchao
tensorboard
ninja
packaging
gradio>=4.44.0
numpy==1.26.4
protobuf==3.20.3
trl
addict